필 Big Decimal 로 데이터 타입 변경, 아이콘 변경

+++ wallet_system/src/main/java/com/wallet/system/vo/InvestmentVO.java
@@ -6,6 +6,7 @@
  */
 package com.wallet.system.vo;
 
+import java.math.BigDecimal;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import org.springframework.stereotype.Repository;
@@ -20,19 +21,19 @@ public class InvestmentVO {
     private int investment_category_id;
     private Date purchase_date;
     private int purchase_size;
-    private float payout_fil;
-    private float fil_paid_per_tb;
+    private BigDecimal payout_fil;
+    private BigDecimal fil_paid_per_tb;
     private boolean is_getting_paid;
 	private int cateogry_fil_per_tb;
-	private float fil_invested;
+	private BigDecimal fil_invested;
 	
 	
     
-	public float getFil_invested() {
+	public BigDecimal getFil_invested() {
 		return fil_invested;
 	}
 
-	public void setFil_invested(float fil_invested) {
+	public void setFil_invested(BigDecimal fil_invested) {
 		this.fil_invested = fil_invested;
 	}
 
@@ -44,19 +45,19 @@ public class InvestmentVO {
 		this.cateogry_fil_per_tb = cateogry_fil_per_tb;
 	}
 
-	public float getFil_paid_per_tb() {
+	public BigDecimal getFil_paid_per_tb() {
 		return fil_paid_per_tb;
 	}
 
-	public void setFil_paid_per_tb(float fil_paid_per_tb) {
+	public void setFil_paid_per_tb(BigDecimal fil_paid_per_tb) {
 		this.fil_paid_per_tb = fil_paid_per_tb;
 	}
 
-	public float getPayout_fil() {
+	public BigDecimal getPayout_fil() {
 		return payout_fil;
 	}
 
-	public void setPayout_fil(float payout_fil) {
+	public void setPayout_fil(BigDecimal payout_fil) {
 		this.payout_fil = payout_fil;
 	}
